Split Rail.tsx (1364 lines) into focused modules (#69)
Pure extraction — no logic or behavioral changes. The monolithic
Rail.tsx is replaced by four focused files under src/views/pages/rail/:

- data.ts: constants (LINE_COLORS, LINE_STATIONS, STATION_COORDS,
  STATION_ORDER), types (StationPill, StationRow), and helpers
  (formatTime, normalizeStation, buildStationRows)
- styles.ts: railStyles() and trainTimelineStyles()
- client.ts: buildInlineJS() inline polling/starred/nearby JS
- index.tsx: all JSX components, re-exports everything for backward
  compatibility

The only import consumer (src/routes/rail.tsx) is updated to point
at the new rail/index module.

Closes #69
